backend/requirements.txt
fastapi>=0.115.0
bcrypt>=4.0.0
uvicorn>=0.30.0
python-dotenv>=1.0.0
openai>=1.30.0
pymupdf>=1.24.0
PyPDF2>=3.0.1
pyyaml>=6.0.2
Pillow>=10.0.0
tqdm>=4.66.0
pydantic>=2.0.0
python-multipart>=0.0.9
agno[postgres]>=2.5.0
sqlalchemy>=2.0.0
psycopg[binary,pool]>=3.1.0
python-docx>=1.1.0
openpyxl>=3.1.0
mammoth>=1.6.0
slowapi>=0.1.9
